backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
openai==1.3.0
httpx==0.25.2
gunicorn==21.2.0
python-jose==3.3.0
requests==2.31.0 